Crate sentry_panic

The Sentry Panic handler integration.

The PanicIntegration, which is enabled by default in sentry, installs a panic handler that will automatically dispatch all errors to Sentry that are caused by a panic. Additionally, panics are forwarded to the previously registered panic hook.

Configuration

The panic integration can be configured with an additional extractor, which might optionally create a sentry Event out of a PanicInfo.

let integration = sentry_panic::PanicIntegration::default().add_extractor(|info| None);

Structs

PanicIntegration

The Sentry Panic handler Integration.

Functions

message_from_panic_info

Extract the message of a panic.

panic_handler

A panic handler that sends to Sentry.